feat: role permission

+++ app/Http/Controllers/ReportController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Laporan;
-use Illuminate\Container\Attributes\Auth;
+use Illuminate\Support\Facades\Auth; // Make sure to use this import
 use Maatwebsite\Excel\Facades\Excel;
 use App\Exports\LaporanExport;
 use Illuminate\Support\Facades\Schema;
@@ -73,6 +73,9 @@ class ReportController extends Controller
      */
     public function edit(string $id)
     {
+        if (!Auth::check() || Auth::user()->role !== 'admin') {
+            return redirect()->route('report')->with('error', 'Anda tidak memiliki akses untuk mengedit laporan.');
+        };
         $laporan = Laporan::findOrFail($id);
         return view('report.edit', compact('laporan'));
     }
@@ -82,6 +85,9 @@ class ReportController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        if (!Auth::check() || Auth::user()->role !== 'admin') {
+            return redirect()->route('report')->with('error', 'Anda tidak memiliki akses untuk mengubah laporan.');
+        };
         $request->validate([
             'nama_teknisi' => 'required',
             'keterangan_kerusakan' => 'required',
@@ -119,6 +125,9 @@ class ReportController extends Controller
      */
     public function destroy(string $id)
     {
+        if (!Auth::check() || Auth::user()->role !== 'admin') {
+            return redirect()->route('report')->with('error', 'Anda tidak memiliki akses untuk menghapus laporan.');
+        };
         $laporan = Laporan::findOrFail($id);
         $laporan->delete();
 
@@ -127,6 +136,9 @@ class ReportController extends Controller
 
     public function bulkDelete(Request $request)
     {
+        if (!Auth::check() || Auth::user()->role !== 'admin') {
+            return redirect()->route('report')->with('error', 'Anda tidak memiliki akses untuk menghapus laporan.');
+        };
         try {
             $ids = $request->input('ids', []);
             if (empty($ids)) {
